+++ hypnotoad/core/mesh.py
@@ -3421,7 +3421,16 @@ class BoutMesh(Mesh):
     def writeArray(self, name, array, f):
         f.write(name, BoutArray(array.centre, attributes=array.attributes))
         f.write(
+            name + "_xlow", BoutArray(array.xlow[:-1, :], attributes=array.attributes)
+        )
+        f.write(
             name + "_ylow", BoutArray(array.ylow[:, :-1], attributes=array.attributes)
+        )
+
+    def writeCorners(self, name, array, f):
+        f.write(
+            name + "_corners",
+            BoutArray(array.corners[:-1, :-1], attributes=array.attributes),
         )
 
     def writeArrayXDirection(self, name, array, f):
@@ -3454,6 +3463,10 @@ class BoutMesh(Mesh):
             # write the 2d fields
             for name in self.fields_to_output:
                 self.writeArray(name, self.__dict__[name], f)
+
+            # write corner positions, as these may be useful for plotting, etc.
+            for name in ["Rxy", "Zxy"]:
+                self.writeCorners(name, self.__dict__[name], f)
 
             if self.user_options.orthogonal:
                 # Also write hy as "hthe" for backward compatibility